Final_Project/users/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
datetime
db-sqlite3
numpy
requests
ipython
pika
kazoo
